==> project/00_Introduction.Rmd <==
---
title: "00_Introduction"
author: "Julio Eduardo Cáceres Pajuelo"
date: '2025-06-27'
output: html_document
editor_options: 
  markdown: 
    wrap: sentence
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Chromatin Accessibility in Spermatogonial Stem Cells During Maturation and Adulthood in a Mouse Model of Early Life Stress

## Introduction

Adverse experiences in early life are risk factors for the development of behavioral and physiological symptoms that can lead to psychiatric and cognitive disorders later in life (1-4).
Some of these symptoms can even be transmitted to the offspring, in some cases through non-genomic mechanisms involving germ cells (3,5).
To investigate the molecular basis of this phenomenon, we analyze chromatin accessibility data generated from a mouse model of unpredictable maternal separation and maternal stress, which mimics early life adversity (2,3).

The main question is whether changes in the epigenome induced by early life stress persist into adulthood.
To address this, we focus on chromatin accessibility in spermatogonial stem cells at two developmental stages: postnatal day 15, shortly after exposure to stress, and in adulthood.
Although we do not directly compare these two time points, analyzing each stage independently allows us to determine whether stress-induced changes in chromatin structure are detectable shortly after exposure and whether similar alterations are still present later in life.
This approach enables us to evaluate the persistence of epigenetic changes and their potential role as a form of chromatin memory in the germline.

## References

1.  Gapp K, Bohacek J, Grossmann J, Brunner AM, Manuella F, Nanni P, Mansuy IM. Potential of environmental enrichment to prevent transgenerational effects of paternal trauma.
    Neuropsychopharmacology.
    2016;41(11):2749--58.
    <doi:10.1038/npp.2016.87>

2.  Smith KE, Pollak SD. Early life stress and development: potential mechanisms for adverse outcomes.
    J Neurodev Disord.
    2020;12(1):34.
    <doi:10.1186/s11689-020-09337-y>

3.  Franklin TB, Russig H, Weiss IC, Gräff J, Linder N, Michalon A, Vizi S, Mansuy IM. Epigenetic transmission of the impact of early stress across generations.
    Biol Psychiatry.
    2010;68(5):408--15.
    <doi:10.1016/j.biopsych.2010.05.036>

4.  Boscardin C, Manuella F, Mansuy IM. Paternal transmission of behavioural and metabolic traits induced by postnatal stress to the 5th generation in mice.
    Environ Epigenet.
    2022;8(1):dvac024.
    <doi:10.1093/eep/dvac024>
    
5.   Gapp K, Jawaid A, Sarkies P, Bohacek J, Pelczar P, Prados J, Farinelli L, Miska E, Mansuy IM. Implication of sperm RNAs in transgenerational inheritance of the effects of early trauma in mice. Nat Neurosci. 2014 May;17(5):667–669. doi:10.1038/nn.3695
